requirements.txt
-e git+ssh://git@github.com/UseAllFive/media-portal.git#egg=media_portal
fabric
django-cors-headers
django_smtp_ssl
drfdocs
django-storages
boto
shortuuid
sendgrid-django
jsonfield